[#5]refactor: 함수 분리
