+++ app/route/services/exception/base_custom_exception.py
@@ -1,4 +1,4 @@
-from app.route.exception.enum.error_enum import ErrorEnum
+from app.route.services.exception.enum.error_enum import ErrorEnum
 from starlette import status
